Update Italian relative time for future
Explanation from my colleague:
In Italian, "in" followed by a time expression is translated with "tra", not with "in". 
There's no difference between "in a day" and "in 1 day" if they are both referred to future plans, they should both be translated with "tra" in Italian.
"In" in Italian refers to locations. It can also refer to time expressions, but not in the future.

The current translation is indeed a mistake and it should be changed to

tra un giorno

See more info about the use of "in" and "tra" in Italian here (not a complete grammar, of course, but it can give you the idea):
http://www.italianlanguageguide.com/grammar/prepositions/